requirements.txt
assertpy==1.0
Django==2.2.9
django-filter==2.2.0
django-json-field==0.5.7
django-model-utils==4.0.0
django-redis==4.11.0
django-smtp-ssl==1.0
djangorestframework==3.10.3
djangorestframework-mango==0.1.1
djangorestframework-simplejwt==4.4.0
model-bakery==1.0.2
psycopg2==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.4
pytz==2019.3
redis==3.3.11
six==1.14.0
sqlparse==0.3.0
